Mailer - By Rajdeep Singh Medtiya
Mailer is a python based script by which a person can send mails to multiple persons which are to be included in a CSV file.
The person needs to type in their login credentials for there mailing service and the content they would like to send to the mails which have been included respectively.
Click Enter and Done. 
The mails are sent to the respective mailing id's.